+++ libft/ft_memcmp.c
@@ -6,15 +6,15 @@
 /*   By: gblanco- <gblanco-@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/20 19:49:16 by gblanco-          #+#    #+#             */
-/*   Updated: 2023/03/20 19:52:56 by gblanco-         ###   ########.fr       */
+/*   Updated: 2023/04/10 12:18:20 by gblanco-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_memcmp(const void *ptr1, const void *ptr2, unsigned long num)
+int	ft_memcmp(const void *ptr1, const void *ptr2, size_t num)
 {
-	unsigned long	i;
+	size_t	i;
 
 	i = 0;
 	while (i < num)
